# openapi.json
{"openapi":"3.0.0","info":{"title":"Task Manager API","version":"1.0.0","description":"REST API for task management with JWT authentication"},"servers":[{"url":"http://localhost:3000","description":"Development server"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"User":{"type":"object","properties":{"id":{"type":"integer"},"username":{"type":"string"},"email":{"type":"string"},"created_at":{"type":"string","format":"date-time"}}},"Task":{"type":"object","properties":{"id":{"type":"integer"},"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string","enum":["pending","in_progress","completed"]},"user_id":{"type":"integer"},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"}}},"ValidationError":{"type":"object","properties":{"errors":{"type":"array","items":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"}}}}}}}},"security":[{"bearerAuth":[]}],"paths":{"/api/tasks":{"get":{"summary":"Get all tasks for authenticated user","tags":["Tasks"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of tasks"}}},"post":{"summary":"Create a new task","tags":["Tasks"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["title"],"properties":{"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string","enum":["pending","in_progress","completed"]}}}}}},"responses":{"201":{"description":"Task created"}}}},"/api/tasks/{id}":{"get":{"summary":"Get task by ID","tags":["Tasks"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Task found"},"404":{"description":"Not found"}}},"put":{"summary":"Update a task","tags":["Tasks"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string","enum":["pending","in_progress","completed"]}}}}}},"responses":{"200":{"description":"Task updated"}}},"delete":{"summary":"Delete a task","tags":["Tasks"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Task deleted"}}}},"/api/auth/register":{"post":{"summary":"Register a new user","tags":["Auth"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["username","email","password"],"properties":{"username":{"type":"string","minLength":3},"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":6}}}}}},"responses":{"201":{"description":"User registered"},"400":{"description":"Validation error"},"409":{"description":"Username or email already exists"}}}},"/api/auth/login":{"post":{"summary":"Login user and get JWT token","tags":["Auth"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"Login successful"},"401":{"description":"Invalid credentials"}}}}},"tags":[]}